indexmain/serlializers.py
from django.db.models import fields
from rest_framework import serializers
from .models import Road


class RoadSerializer(serializers.ModelSerializer):
    class Meta:
        model= Road
        fields= '__all__'