requirements.txt
arrow==1.2.1
asgiref==3.3.4
binaryornot==0.4.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
cookiecutter==1.7.3
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
distlib==0.3.4
Django==3.2.2
django-ckeditor==6.1.0
django-import-export==2.7.1
django-js-asset==1.2.2
djangorestframework==3.12.4
et-xmlfile==1.1.0
filelock==3.4.2
gunicorn==20.1.0
idna==3.3
Jinja2==3.0.3
jinja2-time==0.2.0
MarkupPy==1.14
MarkupSafe==2.0.1
mysqlclient==2.0.3
odfpy==1.4.1
openpyxl==3.0.7
packaging==21.3
pandas==1.2.4
Pillow==8.2.0
platformdirs==2.4.1
plyer==2.0.0
poyo==0.5.0
psycopg2==2.9.1
pyparsing==3.0.6
python-dateutil==2.8.1
python-decouple==3.4
python-slugify==5.0.2
pytz==2021.1
PyYAML==6.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.1.0
text-unidecode==1.3
urllib3==1.26.7
virtualenv==20.13.0
websocket-client==1.2.1
whitenoise==5.3.0
wrapt==1.13.3
xlrd==2.0.1
xlwt==1.3.0
